move src/pageComponents/home to app/home/_components
