+++ template/web/app/home/_components/Guide.tsx
